chore: cleanup
- create stateless widget for search field and use in home
- create stateless widget for home and use in main
